Split analysis/cache.rs into focused sub-modules (#565)
Decompose the ~1,300-line cache.rs into a cache/ directory with focused
sub-modules, each under ~300 lines:

- cache/mod.rs — RecordCache struct and public API re-exports
- cache/loading_strategy.rs — LoadingStrategy enum, select_loading_strategy
- cache/lru_cache.rs — LruRecordCache, LruCacheStats, eviction logic
- cache/compressed_cache.rs — CompressedLruRecordCache (LZ4)
- cache/tiered_cache.rs — TieredRecordCache, auto strategy selection
- cache/serialisation.rs — Binary serialisation, CompressedCacheEntry

No public API changes. All existing tests pass without modification.
